refactor: replace ssize_t with int and long
- Use constexpr for declaring buffer array size

+++ lib/core/include/core/network/Client.hpp
@@ -22,7 +22,7 @@ public:
   ~Client();
   int connect(const std::string&, int);
   int disconnect(int);
-  ssize_t send(int, const MessageType&, const char*, size_t);
+  int send(int, const MessageType&, const char*, size_t);
   void start();
   void stop();
 };
